file structure fixed?
