--- frontend/src/Components/AddArticle.js
@@ -1,29 +1,0 @@
-import React, { useState } from 'react';
-import { axiosWithAuth } from "../data/axiosAuth";
-
-
-const initialArticle = {
-    title: "",
-    summary: "",
-    link: "",
-    image: img,
-    category: "", 
-};
-
-
-const ArticleList = ({ articles, updateArticles }) => {
-    console.log(articles);
-    const [edit, setEdit] = useState(false);
-    const [articleToEdit, setArticleToEdit] = useState(initialArticle);
-
-
-const editArticle = article => {
-    setEdit(true);
-    setArticleToEdit(color);
-};
-const saveEdit = e => {
-    e.preventDefault();
-};
-
-
-}
