+++ package.json
@@ -3,18 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.25",
-    "@fortawesome/free-solid-svg-icons": "^5.11.2",
-    "@fortawesome/react-fontawesome": "^0.1.6",
     "axios": "^0.19.0",
-    "bootstrap": "^4.3.1",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
-    "react-fontawesome": "^1.6.1",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.2.0",
-    "reactstrap": "^8.0.1",
-    "styled-components": "^4.4.0"
+    "react-scripts": "3.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
